Start working on a documentation generation script
This script takes yaml files as input and produces documentation in
markdown format. The thought behind this is that it allows us to
update documentation more efficiently while also allowing us to
automatically generate API typings for editing platforms (eg. for Tua
or ZeroBrane). It also makes it easy to provide users with a changes
diff in the changelog.